backend/requirements.txt
Django>=5.0
djangorestframework>=3.15
django-cors-headers>=4.0
Pillow>=10.0
pymysql>=1.1
